rename validation testimony and create organization validation
