+++ routes/testimony.js
@@ -2,7 +2,7 @@ const express = require('express')
 const router = express.Router()
 const {
   validatedTestimony
-} = require('../middlewares/validations/testimonyValidations')
+} = require('../middlewares/validations/testimony')
 const postTestimonyController = require('../controllers/testimonyController')
 
 router.get('/', function (req, res, next) {
